+++ Source/WorldOfEmpires/TP_ThirdPerson/TP_ThirdPerson.cpp
@@ -1,9 +1,0 @@
-// Copyright Epic Games, Inc. All Rights Reserved.
-
-#include "WorldOfEmpires.h"
-#include "TP_ThirdPerson.h"
-#include "Modules/ModuleManager.h"
-
-IMPLEMENT_PRIMARY_GAME_MODULE( FDefaultGameModuleImpl, TP_ThirdPerson, "TP_ThirdPerson" );
-
-DEFINE_LOG_CATEGORY(LogTP_ThirdPerson)