+++ Source/WorldOfEmpires/TP_ThirdPerson/TP_ThirdPersonGameMode.cpp
@@ -1,8 +1,0 @@
-// Copyright Epic Games, Inc. All Rights Reserved.
-
-#include "TP_ThirdPersonGameMode.h"
-
-ATP_ThirdPersonGameMode::ATP_ThirdPersonGameMode()
-{
-	// stub
-}